+++ CHANGELOG.md
@@ -19,6 +19,16 @@
 * Known Issues:
   * Drop In view controller is not yet compatible with PayPal Touch (app switch auth)
 
+## 3.1.3 (2014-08-22)
+
+* Fix another PayPal payment method display issue in Drop In UI
+
+## 3.1.2 (2014-08-21)
+
+* Fixes
+  * Minor internationalization issue
+  * PayPal payment method display issue in Drop In UI
+
 ## 3.1.1 (2014-08-17)
 
 * Enhancements